# packages/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alumni API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "\ud1b5\ud569 \ub85c\uadf8\uc778 (MemberAuth \ub2e8\uc77c \uacbd\ub85c).\n\n\ub808\uc774\ud2b8\ub9ac\ubc0b\uc740 \uc774 \ub77c\uc6b0\ud130\uc5d0\uc11c 1\ud68c\ub9cc \ucc28\uac10 (\uc11c\ube44\uc2a4 \ud568\uc218\uc5d0\uc11c\ub294 skip).",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Login Auth Login Post"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout Endpoint",
        "description": "\ub85c\uadf8\uc544\uc6c3 (\uc138\uc158 \uc81c\uac70).",
        "operationId": "logout_endpoint_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "description": "\uad00\ub9ac\uc790 \uc790\uae30\uc815\ubcf4 (\ud558\uc704\ud638\ud658).",
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "integer"
                      }
                    ]
                  },
                  "type": "object",
                  "title": "Response Me Auth Me Get"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/auth/member/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Member Login",
        "description": "\uba64\ubc84 \ub85c\uadf8\uc778.",
        "operationId": "member_login_auth_member_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberLoginPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Member Login Auth Member Login Post"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/auth/member/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Member Signup",
        "description": "\uc2e0\uaddc \uac00\uc785\uc2e0\uccad \uc0dd\uc131.",
        "operationId": "member_signup_auth_member_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupRequestRead"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/auth/member/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Member Logout",
        "description": "\uba64\ubc84 \ub85c\uadf8\uc544\uc6c3.",
        "operationId": "member_logout_auth_member_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/auth/member/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Member Me",
        "description": "\uba64\ubc84 \uc774\uba54\uc77c \uc870\ud68c.",
        "operationId": "member_me_auth_member_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Member Me Auth Member Me Get"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/auth/member/activate": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Member Activate Endpoint",
        "description": "\uba64\ubc84 \ud65c\uc131\ud654 (\ud1a0\ud070 \uac80\uc99d \ud6c4 \uacc4\uc815 \uc0dd\uc131/\ube44\ubc00\ubc88\ud638 \uc124\uc815).",
        "operationId": "member_activate_endpoint_auth_member_activate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberActivatePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Member Activate Endpoint Auth Member Activate Post"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/auth/member/change-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change Password",
        "description": "\ube44\ubc00\ubc88\ud638 \ubcc0\uacbd.",
        "operationId": "change_password_auth_member_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Change Password Auth Member Change Password Post"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
    "/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Session",
        "description": "\ud1b5\ud569 \uc138\uc158 \uc870\ud68c. directory_consent_at\u00b7directory_choice_at\ub97c \ud3ec\ud568\ud55c\ub2e4.",
        "operationId": "session_auth_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Session Auth Session Get"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
...
    "/me/": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Get Me",
        "operationId": "get_me_me__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      },
      "put": {
        "tags": [
          "me"
        ],
        "summary": "Update Me",
        "operationId": "update_me_me__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRead"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "400": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "401": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "429": {
            "$ref": "#/components/responses/ProblemDetailsError"
          },
          "500": {
            "$ref": "#/components/responses/ProblemDetailsError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordPayload": {
        "properties": {
          "current_password": {
            "type": "string",
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordPayload"
      },
...
      "LoginPayload": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "password"
        ],
        "title": "LoginPayload"
      },
      "MemberActivatePayload": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "token",
          "password"
        ],
        "title": "MemberActivatePayload"
      },
...
      "MemberLoginPayload": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "password"
        ],
        "title": "MemberLoginPayload"
      },
      "MemberRead": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cohort": {
            "type": "integer",
            "title": "Cohort"
          },
          "major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major"
          },
          "roles": {
            "type": "string",
            "title": "Roles",
            "default": "member"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "all",
              "cohort",
              "private"
            ],
            "title": "Visibility",
            "default": "private"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "birth_lunar": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Lunar"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "company_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Phone"
          },
          "addr_personal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Addr Personal"
          },
          "addr_company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Addr Company"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "suspended",
              "rejected"
            ],
            "title": "Status",
            "default": "active"
          },
          "directory_consent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directory Consent At"
          },
          "directory_choice_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directory Choice At"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "email",
          "name",
          "cohort",
          "id",
          "avatar_url"
        ],
        "title": "MemberRead"
      },
...
      "SignupRequestCreate": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cohort": {
            "type": "integer",
            "title": "Cohort"
          },
          "major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "email",
          "name",
          "cohort",
          "phone"
        ],
        "title": "SignupRequestCreate"
      },
...
      "SignupRequestRead": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cohort": {
            "type": "integer",
            "title": "Cohort"
          },
          "major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "activated"
            ],
            "title": "Status",
            "default": "pending"
          },
          "requested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "activated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "decided_by_student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By Student Id"
          },
          "reject_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reject Reason"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "email",
          "name",
          "cohort",
          "id"
        ],
        "title": "SignupRequestRead"
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          }
        },
        "required": [
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetailsResponse",
        "type": "object"
      }
    },
    "responses": {
      "ProblemDetailsError": {
        "description": "Problem Details error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    }
  }
}